# doc/api.yaml
openapi: 3.0.0
servers:
  - url: "http://localhost:3000/v1"
info:
  title: WASAText
  description: Fantastic Coffee (decaffeinated), the skeleton project for Web and Software architecture course
  version: 1.0.0
...
paths:
...
  /users:
    get:
      tags: ["Profilo"]
      operationId: searchUsers
      summary: Search WasaText users
      description: |-
        Cerca altri utenti registrati tramite username.
      security:
        - BearerAuth: []
      parameters:
        - name: username
          in: query
          required: false
          schema:
            type: string
            minLength: 1
            maxLength: 16
            pattern: "^.*?$"
          description: Username da cercare
      responses:
        "200":
          description: Lista degli utenti trovati
          content:
            application/json:
              example:
                users:
                  - id: "usr_123"
                    username: "mario_rossi"
                    photoUrl: "https://example.com/photo.jpg"
              schema:
                type: object
                description: Contenitore della lista utenti
                required:
                  - users
                properties:
                  users:
                    type: array
                    description: Elenco degli utenti trovati
                    minItems: 0
                    maxItems: 100
                    items:
                      $ref: "#/components/schemas/User"
        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /user/username:
    put:
      tags: ["Profilo"]
      operationId: setMyUserName
      summary: Change WasaText profile username
      description: |-
        Aggiorna il nome visualizzato dall'utente.
      security:
        - BearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserName"
              
      responses:
        "200":
          description: Successful retrieval of WasaText profile username
          content:
            application/json:
              example:
                username: "mario_nuovo"
              schema:
                $ref: "#/components/schemas/UserName"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "409": { $ref: "#/components/responses/Conflict" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /user/photo:
    put:
      tags: ["Profilo"]
      operationId: setMyPhoto
      summary: Change WasaText profile photo
      description: |-
        Aggiorna la foto profilo dell'utente.
      security:
        - BearerAuth: []
      
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserPhoto"

      responses:
        "200":
          description: Successful retrieval of WasaText profile photo
          content:
            application/json:
              example:
                photoUrl: "https://example.com/new_photo.jpg"
              schema:
                $ref: "#/components/schemas/UserPhoto"

        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }

    get:
      tags: ["Profilo"]
      operationId: getMyPhoto
      summary: Retrieve WasaText profile photo
      description: |-
        Ottiene la foto profilo dell'utente.
      security:
        - BearerAuth: []
      
      responses:
        "200":
          description: Successful retrieval of WasaText profile photo
          content:
            application/json:
              example:
                photoUrl: "https://example.com/photo.jpg"
              schema:
                $ref: "#/components/schemas/UserPhoto"

        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /user:
    get:
      tags: ["Profilo"]
      operationId: getMyUserInfo
      summary: Retrieve WasaText profile information
      description: |-
        Ottiene le informazioni del profilo dell'utente.
      security:
        - BearerAuth: []
      
      responses:
        "200":
          description: Successful retrieval of WasaText profile information
          content:
            application/json:
              example:
                id: "usr_123"
                username: "mario_rossi"
                photoUrl: "https://example.com/photo.jpg"
              schema:
                $ref: "#/components/schemas/User"

        "401": { $ref: "#/components/responses/Unauthorized" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
  /conversations/{conversationId}/messages:
    parameters:
      - $ref: "#/components/parameters/conversationId"
    post:
      tags: ["Messages"]
      operationId: sendMessage
      summary: Send a WasaText message
      description: |-
        Invia un nuovo messaggio di testo o foto in una chat specifica.
      security:
        - BearerAuth: []

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewMessageRequest"
      responses:
        "201":
          description: Message sent successfully
          content:
            application/json:
              example:
                id: "msg_101"
                conversationId: "conv_123"
                senderId: "usr_123"
                senderUsername: "mario_rossi"
                dataSent: "2026-03-30T10:05:00Z"
                status:
                  value: "sent"
                content:
                  text: "Come stai?"
              schema:
                $ref: "#/components/schemas/Message"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /conversations/{conversationId}/messages/{messageId}/forward:
    post:
      tags: ["Messages"]
      operationId: forwardMessage
      summary: Forward a WasaText message
      description: |-
        Inoltra un messaggio di testo o foto a un'altra chat.
      security:
        - BearerAuth: []
      parameters:
        - $ref: "#/components/parameters/conversationId"
        - $ref: "#/components/parameters/messageId"

      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageForward"
      responses:
        "201":
          description: Message forwarded successfully
          content:
            application/json:
              example:
                id: "msg_102"
                conversationId: "conv_456"
                senderId: "usr_123"
                senderUsername: "mario_rossi"
                dataSent: "2026-03-30T10:06:00Z"
                isForwarded: true
                status:
                  value: "sent"
                content:
                  text: "Messaggio inoltrato"
              schema:
                $ref: "#/components/schemas/Message"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /conversations/{conversationId}/messages/{messageId}:
    parameters:
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/messageId"

    delete:
      tags: ["Messages"]
      operationId: deleteMessage
      summary: Delete a WasaText message
      description: |-
        Elimina un messaggio di testo o foto da una chat specifica.
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Message deleted successfully
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /conversations/{conversationId}/messages/{messageId}/reactions:
    parameters:
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/messageId"
    post:
      tags: ["Reactions"]
      operationId: commentMessage
      summary: Add a reaction to a WasaText message
      description: |-
        Aggiunge una reazione (like, love, etc.) a un messaggio di testo o foto.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/NewReaction"

      responses:
        "201":
          description: Reaction added successfully
          content:
            application/json:
              example:
                id: "react_123"
                reactionType: "like"
                userSenderId: "usr_123"
                username: "mario_rossi"
              schema:
                $ref: "#/components/schemas/Reaction"
        "400": { $ref: "#/components/responses/BadRequest" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
  /conversations/{conversationId}/messages/{messageId}/reactions/{reactionId}:
    parameters:
      - $ref: "#/components/parameters/conversationId"
      - $ref: "#/components/parameters/messageId"
      - $ref: "#/components/parameters/reactionId"
    delete:
      tags: ["Reactions"]
      operationId: uncommentMessage
      summary: Remove a reaction from a WasaText message
      description: |-
        Rimuove una reazione (like, love, etc.) da un messaggio di testo o foto.
      security:
        - BearerAuth: []
      responses:
        "204":
          description: Reaction removed successfully
        "401": { $ref: "#/components/responses/Unauthorized" }
        "404": { $ref: "#/components/responses/NotFound" }
        "500": { $ref: "#/components/responses/InternalServerError" }
...
components:
  schemas:
...
    User:
      title: User
      description: Informazioni di base di un utente
      type: object
      required: ['id', 'username']
      properties:
        id:
          type: string
          description: Identificatore univoco utente
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_\\-]+$"
        username:
          type: string
          description: Nome utente univoco
          minLength: 3
          maxLength: 16
          pattern: "^.*?$"
        photoUrl:
          type: string
          description: URL della foto profilo dell'utente (stringa vuota se assente)
          minLength: 0
          maxLength: 2048
          pattern: "^(https?://.+)?$"
    UserName:
      type: object
      title: UserName
      description: New username profile
      required: 
        - username
      properties:
        username:
          type: string
          description: Nuovo username da impostare nel profilo
          minLength: 3
          maxLength: 16
          pattern: "^.*?$"
    UserPhoto:
      type: object
      title: UserPhoto
      description: New photo profile
      required: 
        - photoUrl
      properties:
        photoUrl:
          type: string
          description: URL della nuova foto profilo da impostare (stringa vuota per rimuoverla)
          minLength: 0
          maxLength: 2048
          pattern: "^(https?://.+)?$"
    Message:
      type: object
      title: Message
      description: WasaText conversation messages
      required:
        - id
        - conversationId
        - senderId
        - content
        - senderUsername
        - dataSent
        - status
      properties:
        id:
          type: string
          description: Identificatore univoco del messaggio
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_\\-]+$"
        conversationId:
          type: string
          description: ID della conversazione a cui appartiene il messaggio
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_\\-]+$"
        senderId:
          type: string
          description: ID dell'utente mittente (stabile anche dopo un cambio di username)
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_\\-]+$"
        replyToMessageId:
          type: string
          description: ID del messaggio a cui questo risponde (assente se non è una risposta)
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_\\-]+$"
        isForwarded:
          type: boolean
          description: true se il messaggio è stato inoltrato da un'altra conversazione
        content:
          $ref: "#/components/schemas/MessageContent"
        senderUsername:
          type: string
          description: Username del mittente del messaggio
          minLength: 3
          maxLength: 16
          pattern: "^.*?$"
        senderPhotoUrl:
          type: string
          description: Foto profilo del mittente (stringa vuota se assente)
          minLength: 0
          maxLength: 2048
          pattern: "^(/media/[a-zA-Z0-9_.\\-]+|https?://.+)?$"
        dataSent:
          type: string
          description: Data e ora di invio del messaggio
          format: date-time
          minLength: 20
          maxLength: 35
          pattern: "^\\d{4}-\\d{2}-\\d{2}T\\d{2}:\\d{2}:\\d{2}"
        status:
          $ref: "#/components/schemas/StatusMessage"
        reaction:
          type: array
          description: Lista delle reazioni associate al messaggio
          minItems: 0
          maxItems: 100
          items:
            $ref: "#/components/schemas/Reaction"
    StatusMessage:
      type: object
      title: StatusMessage
      description: Descrive lo stato di consegna e lettura del messaggio
      required:
        - value
      properties:
        value:
          type: string
          description: Stato corrente del messaggio
          enum:
            - sent
            - delivered
            - read
            - failed
    MessageContent:
      title: Message Content
      description: |-
        Contenuto di un messaggio: testo, immagine o entrambi.
        Almeno uno dei due campi deve essere valorizzato.
      type: object
      properties:
        text:
          type: string
          description: Testo del messaggio (stringa vuota se il messaggio è solo immagine)
          minLength: 0
          maxLength: 2000
          pattern: "^[\\s\\S]*$"
          example: Ciao a tutti!
        photoUrl:
          type: string
          description: |-
            URL dell'immagine: relativo per i file caricati con uploadMedia
            (es. /media/med_xxx.png) oppure assoluto. Stringa vuota se assente.
          minLength: 0
          maxLength: 2048
          pattern: "^(/media/[a-zA-Z0-9_.\\-]+|https?://.+)?$"
...
    MessageForward:
      title: Message Forward
      description: WasaText message forwarding
      type: object
      required: 
        - targetConversationId
      properties:
        targetConversationId:
          type: string
          description: ID della conversazione di destinazione dell'inoltro
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_\\-]+$"
          example: "conv_67890"
    NewReaction:
      title: New Reaction
      description: |-
        Payload inviato dal client per aggiungere una reazione a un messaggio.
        L'utente che reagisce NON viene indicato qui: è ricavato dal token
        presente nell'header Authorization.
      type: object
      required:
        - reactionType
      properties:
        reactionType:
          type: string
          description: Tipo/Emoji di reazione utilizzata
          minLength: 1
          maxLength: 20
          pattern: "^.+$"
          example: "like"
    Reaction:
      title: Reaction
      description: Reazione applicata a un messaggio, così come restituita dal server
      type: object
      required:
        - id
        - reactionType
        - userSenderId
      properties:
        id:
          type: string
          description: ID della reazione, usato dal client per rimuoverla (uncommentMessage)
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_\\-]+$"
        reactionType:
          type: string
          description: Tipo/Emoji di reazione utilizzata
          minLength: 1
          maxLength: 20
          pattern: "^.+$"
          example: "like"
        userSenderId:
          type: string
          description: ID dell'utente che ha aggiunto la reazione
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_\\-]+$"
          example: "user_67890"
        username:
          type: string
          description: Username dell'utente che ha aggiunto la reazione
          minLength: 3
          maxLength: 16
          pattern: "^.*?$"
...
    NewMessageRequest:
      title: New Message Request
      description: Payload per inviare un nuovo messaggio in una conversazione
      type: object
      required: 
        - content
      properties:
        content:
          $ref: "#/components/schemas/MessageContent"
        replyToMessageId:
          type: string
          description: ID of the message being replied to
          minLength: 1
          maxLength: 64
          pattern: "^[a-zA-Z0-9_\\-]+$"
    ErrorResponse:
      title: Error Response
      description: Error response returned by the server in case of failure
      type: object
      required:
        - message
      properties:
        message:
          type: string
          description: Descrizione dell'errore
          example: "Risorsa non trovata"
...
  responses:
    Unauthorized:
      description: The access token is missing or it's expired
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    BadRequest:
      description: The request was not compliant with the documentation (eg. missing fields, etc)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    NotFound:
      description: The requested resource (conversation, message, group, reaction, user) does not exist
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    Conflict:
      description: The request could not be completed due to a conflict with the current state of the resource (e.g. username already taken)
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
    InternalServerError:
      description: The server encountered an internal error. Further info in server logs
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      description: |-
        Identificatore utente semplificato (non un JWT reale) ottenuto dall'operazione 'doLogin'.
        Deve essere incluso nell'header 'Authorization' per tutte le richieste protette, nel formato 'Bearer <identifier>'.
  parameters:
    conversationId:
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[a-zA-Z0-9_\\-]+$"
      description: The ID of the conversation to retrieve messages from
      name: conversationId
      in: path
      required: true
...
    messageId:
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[a-zA-Z0-9_\\-]+$"
      name: messageId
      in: path
      required: true
      description: The ID of the message
    reactionId:
      schema:
        type: string
        minLength: 1
        maxLength: 64
        pattern: "^[a-zA-Z0-9_\\-]+$"
      name: reactionId
      in: path
      required: true
      description: The ID of the reaction to remove